feat(symcache) changed types for one-per-dsym cache
